Refactor ESQL query
